changed file structure and added error handling
