+++ frontend/src/App.js
@@ -1,25 +1,25 @@
 import React from "react";
-import Register from "./pages/register";
 import { Route, Routes } from "react-router-dom";
-import Container from "react-bootstrap/Container";
-import Login from "./pages/login";
-import Dashboard from "./pages/dashboard";
-import Profile from "./pages/profile";
-import EditProfile from "./pages/editProfile";
-import Follows from "./pages/follows";
-import NavBar from "./components/navbar";
-import AdminCategories from "./pages/adminCategories";
 import { UserContextProvider } from "./utils/userContext";
-import EditCategory from "./pages/editCategory";
-import AddCategory from "./pages/addCategory";
-import UserList from "./pages/usersList";
-import Categories from "./pages/categories";
-import AdminQuestions from "./pages/adminQuestions";
-import LessonQuiz from "./pages/lesson";
-import AdminDashboard from "./pages/adminDashboard";
-import Result from "./pages/result";
 import { LessonContextProvider } from "./utils/lessonContext";
-import AdminAddWord from "./pages/adminAddWord";
+import Container from "react-bootstrap/Container";
+import NavBar from "./components/navbar";
+import Register from "./pages/register";
+import Login from "./pages/login";
+import Dashboard from "./pages/user/dashboard";
+import Profile from "./pages/user/profile";
+import EditProfile from "./pages/user/editProfile";
+import Follows from "./pages/user/follows";
+import Categories from "./pages/user/categories";
+import LessonQuiz from "./pages/user/lesson";
+import Result from "./pages/user/result";
+import UserList from "./pages/usersList";
+import AdminCategories from "./pages/admin/categories";
+import EditCategory from "./pages/admin/editCategory";
+import AddCategory from "./pages/admin/addCategory";
+import AdminQuestions from "./pages/admin/questions";
+import AdminDashboard from "./pages/admin/dashboard";
+import AdminAddWord from "./pages/admin/addWord";
 
 function App() {
   return (
@@ -32,14 +32,9 @@ function App() {
             <Route path="/login" element={<Login />} />
             <Route path="/home" element={<Dashboard />} />
             <Route path="/profile/:id" element={<Profile />} />
-            <Route path="/profile" element={<Profile />} />
             <Route path="/profile/edit" element={<EditProfile />} />
             <Route path="/profile/follows" element={<Follows />} />
-            <Route path="admin/home" element={<AdminDashboard />} />
-            <Route path="/admin/categories" element={<AdminCategories />} />
-            <Route path="/admin/category/:id/edit" element={<EditCategory />} />
-            <Route path="/admin/category/add" element={<AddCategory />} />
-            <Route path="/users" element={<UserList />} />
+            <Route path="/profile" element={<Profile />} />
             <Route path="/categories" element={<Categories />} />
             <Route path="/lesson/:id/results" element={<Result />} />
             <Route
@@ -50,12 +45,17 @@ function App() {
                 </LessonContextProvider>
               }
             />
+            <Route path="/users" element={<UserList />} />
+            <Route path="admin/home" element={<AdminDashboard />} />
+            <Route path="/admin/categories" element={<AdminCategories />} />
+            <Route path="/admin/category/:id/edit" element={<EditCategory />} />
+            <Route path="/admin/category/add" element={<AddCategory />} />
             <Route
               path="/admin/category/:id/questions"
               element={<AdminQuestions />}
             />
             <Route
-              path="/admin/category/:id/addWord"
+              path="/admin/category/:id/question/add"
               element={<AdminAddWord />}
             />
           </Routes>
