+++ frontend/src/pages/user/categories.js
@@ -3,8 +3,8 @@ import { useNavigate } from "react-router-dom";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import CardGroup from "react-bootstrap/CardGroup";
-import CategoryCard from "../components/categoryCard";
-import * as lessonService from "../services/lessonService";
+import CategoryCard from "../../components/categoryCard";
+import * as lessonService from "../../services/lessonService";
 
 const Categories = () => {
   const [categories, setCategories] = useState();
