+++ frontend/src/pages/user/dashboard.js
@@ -1,16 +1,13 @@
-import React, { useContext } from "react";
+import React from "react";
 import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import Image from "react-bootstrap/Image";
-import { UserContext } from "../utils/userContext";
-import WordsPanel from "../components/learnedWordsPanel";
-import avatarPlaceholder from "../asset/avatar_placeholder.png";
+import WordsPanel from "../../components/learnedWordsPanel";
+import avatarPlaceholder from "../../asset/avatar_placeholder.png";
 import { Link } from "react-router-dom";
 
 const Dashboard = () => {
-  const { user } = useContext(UserContext);
-
   return (
     <Container>
       <h1>Dashboard</h1>
